libra_downstream/requirements.txt
torch>=2.2.0
torchvision>=0.17.0
pycocotools>=2.0.7
tqdm>=4.66.0
numpy>=1.24.0
pillow>=10.0.0
thop>=0.1.1.post2209072238
fvcore>=0.1.5.post20221221
opencv-python>=4.8.0
scipy>=1.10.0
